+++ pkg/deployment/features/features.go
@@ -22,6 +22,11 @@ package features
 
 import "github.com/arangodb/go-driver"
 
+const (
+	Enabled  = "true"
+	Disabled = "false"
+)
+
 var _ Feature = &feature{}
 
 type Feature interface {